Better accessibility for main args from sub commands
Fixes #40 and Fixes #11 (in jontejj).
