+++ jargo/src/test/java/se/softhouse/jargo/ArityArgumentTest.java
@@ -27,7 +27,6 @@ import java.util.List;
 import org.junit.Test;
 
 import se.softhouse.common.testlib.Explanation;
-import se.softhouse.jargo.CommandLineParserInstance.ArgumentIterator;
 import se.softhouse.jargo.internal.Texts.ProgrammaticErrors;
 import se.softhouse.jargo.internal.Texts.UserErrors;
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
