+++ jargo/src/test/java/se/softhouse/jargo/commands/Git.java
@@ -1,0 +1,39 @@
+/* Copyright 2018 jonatanjonsson
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+package se.softhouse.jargo.commands;
+
+import se.softhouse.jargo.Command;
+import se.softhouse.jargo.ParsedArguments;
+import se.softhouse.jargo.commands.Commit.Repository;
+
+public class Git extends Command
+{
+	public Git(Repository repo)
+	{
+		super(subCommands(new Commit(repo), new Log(repo)));
+	}
+
+	@Override
+	protected void execute(ParsedArguments parsedArguments)
+	{
+
+	}
+
+	@Override
+	protected String commandName()
+	{
+		return "git";
+	}
+}
